+++ provider/pkg/gen/schema.go
@@ -618,11 +618,9 @@ additional information about using Server-Side Apply to manage Kubernetes resour
 
 	nodeDeps := map[string]string{
 		"shell-quote":       "^1.6.1",
-		"tmp":               "^0.0.33",
-		"@types/tmp":        "^0.0.33",
 		"glob":              "^10.3.10",
-		"node-fetch":        "^2.3.0",
-		"@types/node-fetch": "^2.1.4",
+		"node-fetch":        "^3.3.2",
+		"@types/node-fetch": "^2.6.11",
 	}
 	if gen.pulumiKubernetesDependency != "" {
 		nodeDeps["@pulumi/kubernetes"] = gen.pulumiKubernetesDependency